+++ circuits/src/ecc/foreign/mod.rs
@@ -12,8 +12,8 @@
 // limitations under the License.
 
 //! Module that implements foreign field EC arithmetic
-pub mod edwards_chip;
-pub(crate) mod gates;
-pub mod weierstrass_chip;
 
-pub use weierstrass_chip::*;
+mod gates;
+
+pub mod edwards_chip;
+pub mod weierstrass_chip;