+++ circuits/src/ecc/native/edwards_chip.rs
@@ -305,10 +305,10 @@ impl EccConfig {
             let one = Expression::from(1);
             let edwards_d = Expression::Constant(C::D);
 
-            let x2 = x.square();
-            let y2 = y.square();
+            let x_sq = x.square();
+            let y_sq = y.square();
 
-            let id = y2.clone() - x2.clone() - (one + edwards_d * x2 * y2);
+            let id = y_sq.clone() - x_sq.clone() - (one + edwards_d * x_sq * y_sq);
 
             Constraints::with_selector(*q_point, vec![("curve equation", id)])
         })
@@ -1134,7 +1134,7 @@ mod tests {
     ecc_tests!(test_msm);
     ecc_tests!(test_msm_by_bounded_scalars);
     ecc_tests!(test_mul_by_constant);
-    ecc_tests!(test_coordinates_edwards);
+    ecc_tests!(test_coordinates);
 
     #[test]
     fn test_htc() {